+++ examples/demos/synergy_infrastructure_removing_network.tf
@@ -1,25 +1,31 @@
+
 provider "oneview" {
-  ov_username =   "${var.username}"
-  ov_password =   "${var.password}"
-  ov_endpoint =   "${var.endpoint}"
-  ov_sslverify =  "${var.ssl_enabled}"
-  ov_apiversion = 2200
-  ov_ifmatch = "*"
+	ov_username = <ov-username>
+	ov_password = <ov-password>
+	ov_endpoint = <ov-endpoint>
+	ov_sslverify = false
+	ov_apiversion = <ov-apiversion>
 }
 
-/*  Below example fetches ethernet network which is a part of LIG, and do a PUT call to update the logical interconnect group, 
-    thereby removing other networks which isnt part of PUT request but exist in LIG.
+/*
+	While executing the following script follow the below steps
+	1. Use target option to update/delete the specific resource. Otherwise there is chance of deleting the Network and NetworkSet first and then updating the LIG and LI.
+	2. To update the LIG and LI follow the below commands
+		terraform apply -target=oneview_logical_interconnect_group.logical_interconnect_group -target=oneview_logical_interconnect.logical_interconnect
+	3. Once the LIG and LI are updated, delete the NetworkSet and Network in order using the below commands.
+		terraform destroy -target=oneview_network_set.network_et
+		terraform destroy -target=oneview_ethernet_network.ethernet_network	
 */
 
-# GET THE ETHERNET NETWORK TO GET THE URI TO ASSIGN TO UPLINKSET
+/* GET THE ETHERNET NETWORK TO GET THE URI TO ASSIGN TO UPLINKSET*/
 data "oneview_ethernet_network" "eth_net" {
-        name = "<network_name>"
+        name = "TestEth"
 }
 
-# REMOVING THE NETWORK FROM LIG 
+/* REMOVING THE NETWORK FROM LIG */
 resource "oneview_logical_interconnect_group" "logical_interconnect_group" {
-	type = "logical-interconnect-groupV8"
-	name = "<LIG_name>"
+	type = "logical-interconnect-groupV5"
+	name = "LIG"
 	interconnect_bay_set = 3
 	enclosure_indexes = [1, 2, 3]
 	redundancy_type = "HighlyAvailable"
@@ -66,25 +72,8 @@ resource "oneview_logical_interconnect_group" "logical_interconnect_group" {
 	}]
 }
 
-# PERFORMING UPDATE FROM GROUP ON LOGICAL INTERCONNECT TO BRING BACK IT TO CONSISTENT STATE 
-/*
-	While executing the following script follow the below steps
-	1. Use target option to update/delete the specific resource. Otherwise there is chance of deleting the Network and NetworkSet first and then updating the LIG and LI.
-	2. To update the LIG and LI follow the below commands
-		terraform apply -target=oneview_logical_interconnect_group.logical_interconnect_group -target=oneview_logical_interconnect.logical_interconnect
-	3. Once the LIG and LI are updated, delete the NetworkSet and Network in order using the below commands.
-		terraform destroy -target=oneview_network_set.network_et
-		terraform destroy -target=oneview_ethernet_network.ethernet_network	
-*/
-
-/*
-resource "oneview_logical_interconnect" "logical_interconnect"{
-}
-*/
-
-/*
+/* PERFORMING UPDATE FROM GROUP ON LOGICAL INTERCONNECT TO BRING BACK IT TO CONSISTENT STATE */
 resource "oneview_logical_interconnect" "logical_interconnect" {
 	update_type = "updateComplianceById"
 	depends_on = ["oneview_logical_interconnect_group.logical_interconnect_group"]
 }
-*/