+++ examples/hypervisor_managers/data_source.tf
@@ -10,7 +10,7 @@ provider "oneview" {
 variable "hm_endpoint" {
   type        = string
   description = "Hypervisor Manager IP"
-  default     = "172.18.13.11" #"<HM-IP>"
+  default     = "<HM-IP>"
 }
 
 # Testing data source